refactor(gpu_hash): split blake2s into concern submodules + internal cleanups
Structure:
- blake2s/mod.rs (779 lines) -> thin facade re-exporting hash + merkle +
  gather + transcript flat; downstream blake2s::<item> paths unchanged
  (the facade re-export is a glob).
- blake2s/hash.rs: leaf-hashing kernels + launchers.
- blake2s/merkle.rs: node kernels + tree builders; the single-coset
  build_merkle_tree is now #[doc(hidden)] (test-reference reader for
  downstream parity tests, alongside the gather trio).
- gather.rs: gains query_index_to_tree_index and the three #[doc(hidden)]
  test-reference readers; a module-doc comment documents the slab layout
  and the test-reference status.

Internal cleanups (no behavior change):
- Merge the two near-identical multi-coset node launchers
  (_separate/_at_offsets) via a shared raw-pointer tail; the two entry
  points remain only to satisfy the differing borrow shapes.
- Factor the repeated inactive-desc-slot validation into
  Oracle{Gather,PartialPath}Desc::assert_inactive.
- Replace the repeated `assert(len <= u32::MAX); len as u32` idiom with a
  crate-private checked_u32 helper.
- Move the runtime WARP_SIZE % STATE_SIZE assert to a crate-level const _.
- Demote never-externally-referenced items to private/pub(crate):
  launch_leaves_kernel, launch_nodes_kernel, build_merkle_tree_nodes_multi_coset,
  Gpu{ChunkedInput,GatherTreeCaps,GatherEAddresses}Desc, and their
  GKR_*_MAX_* constants.

Gates: cargo check -p gpu_hash --all-targets (warning-clean),
cargo check -p gpu_circuit_prover, gpu_hash release suite 29/29 under GPU lock.
